% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_meta_data_where_possible.R
\name{compare_meta_data_where_possible}
\alias{compare_meta_data_where_possible}
\title{Compare meta data where possible.}
\usage{
compare_meta_data_where_possible(data1, data2, suffix = NULL)
}
\arguments{
\item{data1}{First \code{GADSdat} object.}

\item{data2}{Second \code{GADSdat} object.}

\item{suffix}{Suffix that should be removed from all}
}
\value{
Returns a named list via \code{eatFDZ::compare_data()}:
\itemize{
\item \code{compared_variables}: Which variables have been compared?
\item \code{differences_variable_labels}: Differences on variable labels.
\item \code{differences_value_labels}: Differences on value labels and missing tags.
}
}
\description{
This function allows comparing the meta data of two data sets for all variables with identical names in both data sets.
}
\details{
If \code{compare_meta_data_where_possible()} is used within data tests, it is highly recommended to test
\code{compared_variables} to ensure that tests are actually passing and not just lacking.
}
\examples{
## tbd

}
