% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_yaml.R
\name{create_data_yaml}
\alias{create_data_yaml}
\title{Create Data Yaml.}
\usage{
create_data_yaml(
  .path = getwd(),
  name,
  release_path,
  oldrel_path = NULL,
  version = "v1.0"
)
}
\arguments{
\item{.path}{Path to the \code{eatDaT} repository. Defaults to the current working directory.}

\item{name}{Name of the data set. The file will be named accordingly.}

\item{release_path}{Path to the data set.}

\item{oldrel_path}{Optional: Path to the previous data version. Defaults to \code{NULL}.}

\item{version}{Version number of the current data set. Defaults to \code{'v1.0'}.}
}
\value{
Creates a .yaml. Returns \code{NULL}.
}
\description{
Create a data yaml that contains paths to the current data version (\code{'release_path'}),
the previous version (\code{'oldrel_path'})
and the current version number (\code{'version'}).
}
\examples{
## tbd

}
