% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_diff.R
\name{create_diff}
\alias{create_diff}
\alias{create_meta_diff}
\alias{create_data_diff}
\title{Create Data and Meta Data Diff.}
\usage{
create_diff(.path = getwd(), name, id, tolerance = sqrt(.Machine$double.eps))

create_meta_diff(.path = getwd(), name)

create_data_diff(
  .path = getwd(),
  name,
  id,
  tolerance = sqrt(.Machine$double.eps)
)
}
\arguments{
\item{.path}{Path to the \code{eatDataTest} repository. Defaults to the current working directory.}

\item{name}{Name of the data set. The file will be named accordingly.}

\item{id}{Name of the id variable(s) in both data set.}

\item{tolerance}{A numeric value greater than or equal to 0. Differences smaller than \code{tolerance} are not reported. The default value is close to 1.5e-8.}
}
\value{
Creates two \code{.xlsx} files. Returns \code{NULL}.
}
\description{
\code{create_meta_diff} creates a list of differing variables as well as lists of meta differences
on variable and on value level.
\code{create_data_diff} create a list of differing variables and a cross tables
for each differing variable.
\code{create_diff} calls both functions.
}
\examples{
## tbd

}
