% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data}
\alias{import_data}
\title{Import Data.}
\usage{
import_data(.path = getwd(), name, data_version = c("release", "oldrel"))
}
\arguments{
\item{.path}{Path to the \code{eatDataTest} repository. Defaults to the current working directory.}

\item{name}{Name of the data set. The file will be named accordingly.}

\item{data_version}{Available options are \code{'release'} and \code{'oldrel'}. Defaults to \code{'release'}.}
}
\value{
Returns a data set in \code{GADSdat} format.
}
\description{
Imports a data set based on its yaml file. Currently, \code{import_data()} supports
\code{.RDS} and \code{.sav} files.
}
\examples{
## tbd

}
