% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_data_tests.R
\name{run_data_tests}
\alias{run_data_tests}
\title{Run Data Tests.}
\usage{
run_data_tests(.path = getwd(), name)
}
\arguments{
\item{.path}{Path to the \code{eatDaT} repository. Defaults to the current working directory.}

\item{name}{Name of the data set. The file will be named accordingly.}
}
\value{
Creates a test result icon. Returns \code{NULL}.
}
\description{
Run all tests for a data set ...
}
\examples{
## tbd

}
