% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_revdep_tests.R
\name{run_revdep_tests}
\alias{run_revdep_tests}
\title{Run Reverse Depedency Tests.}
\usage{
run_revdep_tests(.path = getwd(), name, overwrite_icons = FALSE)
}
\arguments{
\item{.path}{Path to the \code{eatDataTest} repository. Defaults to the current working directory.}

\item{name}{Name of the data set.}

\item{overwrite_icons}{Should test result icons of the reverse dependenies be overwritten?.}
}
\value{
Returns a test report.
}
\description{
Run all reverse dependency tests for a data set.
}
\details{
If a project contains multiple data sets, these can sometimes depend on each other.
For example, a multiple imputaions data may depend on the unimputed data set.
\code{run_revdep_tests()} borrows the idea of reverse dependency tests from the R package
\code{revdep} and applies it do data sets.

For instance, if a project contains three data sets (A, B, C). Both data sets B and C
depend on data set A. If \code{run_revdep_tests()} is run for data set A,
data tests will be performed for data sets B and C.
}
\examples{
## tbd

}
