% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_data.R
\name{add_data}
\alias{add_data}
\title{Add Data.}
\usage{
add_data(
  .path = getwd(),
  name,
  release_path,
  oldrel_path = "",
  version = "v1.0"
)
}
\arguments{
\item{.path}{Path to the \code{eatDaT} repository. Defaults to the current working directory.}

\item{name}{Name of the data set. All files will be named accordingly.}

\item{release_path}{Path to the data set.}

\item{oldrel_path}{Optional: Path to the previous data version.}

\item{version}{Version number of the current data set. Defaults to \code{'v1.0'}.}
}
\value{
Creates data set infrastructure. Returns a markdown snippet.
}
\description{
Add the infrastructure for a new data set to the \code{eatDataTest} repository. This includes
\itemize{
\item creating an empty \code{GADSdat} object if \code{oldrel} is empty
\item a \code{.yaml} file
\item a markdown changelog (\code{.md} file)
\item a test file (\code{.R} file in \code{tests})
\item an initial test result icon (\code{.svg} file in \code{tests})
\item a snippet which can be inserted into the \code{Readme.md}.
}
}
\examples{
## tbd

}
